requirements.txt
kivy
python-docx
